requirements.txt
# FastAPI Restaurant Order Management System
# Python 3.9+ required

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
pyodbc==5.0.1
sqlalchemy==2.0.23

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Documentation
python-dotenv==1.0.0

# Testing (optional)
pytest==7.4.3
httpx==0.25.1
